aliasmgr: allow persisting manually added aliases
Previously we were able to manually add scid aliases for channels but
that would slightly misbehave especially around persistence and
restarts. By default when a channel gets confirmed it won't get any of
its aliases reloaded in memory. We change the existing base-lookup flag
to also signal that an alias may be forcefully persisted and reloaded to
the in-memory maps when the alias manager starts up.

We achieve this by appending an extra byte to all the alias storage
entries. Previously we'd have 8 bytes, and now we extend the entry to
have 9 bytes. For old entries, the x00 byte is assumed which signals
non-persistency. Any new entries will always have 9 bytes stored with
the last byte serving as the signal for persistence.